requirement.txt
pandas==1.3.5
numpy==1.21.6
sklearn==1.0.2
lightgbm==3.3.2
matplotlib==3.5.1
PyYAML==6.0